refactor: move job schema
