+++ internal/schema/job/affinity.go
