+++ internal/schema/job/artifact.go
