+++ internal/schema/job/network.go
