+++ internal/schema/job/variable.go
