+++ internal/schema/job/variables.go
